[ymeng] use big-integer as result of power function for avoiding beyond the boundary

+++ src/test/java/tw/ymeng/algorithm/proposition/MathExTest.java
@@ -2,6 +2,8 @@ package tw.ymeng.algorithm.proposition;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static tw.ymeng.algorithm.proposition.MathEx.pow;
@@ -10,6 +12,6 @@ public class MathExTest {
 
     @Test
     public void should_get_power() {
-        assertThat(pow(2L, 10), is(1024L));
+        assertThat(pow(2, 10), is(BigInteger.valueOf(1024)));
     }
 }
